Image as a one dimensional vector
Normal image should be a matrix with size M X N. M and N represent the number of pixels per row and column. Your job is to find the M and N of our anatomical input image. From the length of this vector, you can find out what is the total pixel of this image. Then you need to guess what can be the possible combination of M and N if MxN is fixed. You try each pair and find out which pair make sense to you.